refactor: make component names consistent
