+++ src/stories/CardOwnerNameInput.stories.tsx
@@ -1,5 +1,5 @@
 import { Story } from "@storybook/react";
-import OwnerName, { OwnerNameProps } from "../components/OwnerName";
+import OwnerName, { OwnerNameProps } from "../components/CardOwnerNameInput";
 
 export default {
   title: "Component/OwnerName",
