+++ src/stories/CardSecurityCodeInput.stories.tsx
@@ -1,12 +1,12 @@
 import { Story } from "@storybook/react";
-import Cvc, { CvcProps } from "../components/CardSecurityCode";
+import Cvc from "../components/CardSecurityCodeInput";
 
 export default {
   title: "Components/Cvc",
   component: Cvc,
 };
 
-const Template: Story<CvcProps> = (args) => <Cvc {...args} />;
+const Template: Story = (args) => <Cvc {...args} />;
 
 export const Default = Template.bind({});
 Default.args = {};
